Improve C++11 atomics to directly route to wasm opcodes rather than go through an asm.js emulation function. Improve C++11 64-bit atomics test and split out asm.js atomics to a separate file that's not built in Wasm mode.
