+++ system/lib/asmjs_pthreads.symbols
@@ -1,0 +1,31 @@
+         T emscripten_atomic_add_u64
+         T emscripten_atomic_and_u64
+         U emscripten_atomic_cas_u32
+         T emscripten_atomic_cas_u64
+         U emscripten_atomic_exchange_u32
+         T emscripten_atomic_exchange_u64
+         T emscripten_atomic_load_f32
+         T emscripten_atomic_load_f64
+         U emscripten_atomic_load_u32
+         T emscripten_atomic_load_u64
+         T emscripten_atomic_or_u64
+         T emscripten_atomic_store_f32
+         T emscripten_atomic_store_f64
+         U emscripten_atomic_store_u32
+         T emscripten_atomic_store_u64
+         T emscripten_atomic_sub_u64
+         T emscripten_atomic_xor_u64
+         T _emscripten_atomic_fetch_and_add_u64
+         T _emscripten_atomic_fetch_and_and_u64
+         T _emscripten_atomic_fetch_and_or_u64
+         T _emscripten_atomic_fetch_and_sub_u64
+         T _emscripten_atomic_fetch_and_xor_u64
+         T ___atomic_load_8
+         T ___atomic_store_8
+         T ___atomic_exchange_8
+         T ___atomic_compare_exchange_8
+         T ___atomic_fetch_add_8
+         T ___atomic_fetch_sub_8
+         T ___atomic_fetch_and_8
+         T ___atomic_fetch_or_8
+         T ___atomic_fetch_xor_8
