+++ system/lib/pthreads.symbols
@@ -13,24 +13,10 @@
          T __vm_unlock
          W __vm_unlock_impl
          T __wait
-         T _emscripten_atomic_fetch_and_add_u64
-         T _emscripten_atomic_fetch_and_and_u64
-         T _emscripten_atomic_fetch_and_or_u64
-         T _emscripten_atomic_fetch_and_sub_u64
-         T _emscripten_atomic_fetch_and_xor_u64
          T _pthread_getcanceltype
          T _pthread_isduecanceled
          T _pthread_msecs_until
          T ___atomic_is_lock_free
-         T ___atomic_load_8
-         T ___atomic_store_8
-         T ___atomic_exchange_8
-         T ___atomic_compare_exchange_8
-         T ___atomic_fetch_add_8
-         T ___atomic_fetch_sub_8
-         T ___atomic_fetch_and_8
-         T ___atomic_fetch_or_8
-         T ___atomic_fetch_xor_8
          T _llvm_atomic_load_add_i32_p0i32
          T _llvm_memory_barrier
          U accept
@@ -56,23 +42,6 @@
          U dup
          U dup2
          U emscripten_asm_const
-         T emscripten_atomic_add_u64
-         T emscripten_atomic_and_u64
-         U emscripten_atomic_cas_u32
-         T emscripten_atomic_cas_u64
-         U emscripten_atomic_exchange_u32
-         T emscripten_atomic_exchange_u64
-         T emscripten_atomic_load_f32
-         T emscripten_atomic_load_f64
-         U emscripten_atomic_load_u32
-         T emscripten_atomic_load_u64
-         T emscripten_atomic_or_u64
-         T emscripten_atomic_store_f32
-         T emscripten_atomic_store_f64
-         U emscripten_atomic_store_u32
-         T emscripten_atomic_store_u64
-         T emscripten_atomic_sub_u64
-         T emscripten_atomic_xor_u64
          U emscripten_futex_wait
          U emscripten_futex_wake
          U emscripten_get_now
